drm/panel: st7703: Improve the power up/down sequence of the panel
The datasheet specifies that there needs to be 120ms delay after
the sleep out command, not after reset as the driver assumes.

The delay between init commands and the sleep out is not necessary.

The datasheet also specifies that it's better to keep reset asserted
while powering up the supplies, and that IOVCC should be enabled
first.

There also needs to be a delay after enabling the supplies and
before deasserting the reset. The datasheet specifies 1ms after
the supplies reach the required voltage. Use 10-20ms to give the
power supplies some time to reach the required voltage, too.

After issuing SLEEP IN, it's necessary to let the controller complete
the blanking sequence and drain the charge from the panel. (According
to the datasheet.)

This fixes panel initialization failures during resume from sleep
on PinePhone.

Signed-off-by: Ondrej Jirman <megi@xff.cz>