Revert "ALSA: hda: Fix page fault in snd_hda_codec_shutdown()"
This reverts commit 7494e2e6c55ed192f2b91c821fd6832744ba8741.

Which was upstream commit f2bd1c5ae2cb0cf9525c9bffc0038c12dd7e1338.

The patch caused a regression leading to the missing HD-audio device
with ASoC SOF driver.  It was a part of large series and backporting
it alone breaks things while backporting the whole is too intrusive
as stable changes.  And, the issue the patch tries to address is a
corner case, hence it's better to revert.

BugLink: https://bugzilla.kernel.org/show_bug.cgi?id=216613
Signed-off-by: Takashi Iwai <tiwai@suse.de>